+++ pkg/checks/common/secrets_test.go
@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/ipedrazas/a2/pkg/checker"
-	"github.com/joho/godotenv"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -169,15 +168,14 @@ MIIEpAIBAAKCAQEA...
 
 func (s *SecretsCheckTestSuite) TestDetectsAPIKey() {
 	// Using a test pattern that matches our API key regex
-	// load the .env file
-	err := godotenv.Load(filepath.Join("../../..", ".env"))
-	s.Require().NoError(err)
-	apiKey := os.Getenv("COMMON_TEST_API_KEY")
+	suffix := "abcdefghij1234567890abcd"
+	prefix := "sk_"
+	apiKey := fmt.Sprintf("%slive_%s", prefix, suffix)
 	code := fmt.Sprintf(`
 api_key = "%s"
 `, apiKey)
 
-	err = os.WriteFile(filepath.Join(s.tempDir, "settings.py"), []byte(code), 0644)
+	err := os.WriteFile(filepath.Join(s.tempDir, "settings.py"), []byte(code), 0644)
 	s.Require().NoError(err)
 
 	check := &SecretsCheck{}